Updated FAQ based on the compiled questions from the discord server.
This is possible thanks to the general for compiling the initial draft, me prof reading and adding notes. The document was also read by the devs of the project and, meeth for delivering me with the final draft that I further prof read. This is the first iteration of the FAQ, further prof reading is advised.